Apply or Set Bounding‐Box Header on an OSM File

Writes a new PBF file (or overwrites it) with its header bounding box set to either a user‐supplied sf::st_bbox or—by default—the computed data bounding box. Computing may require reading the entire file and can be slow!

Usage

osm_set_bbox(
  input_path,
  output_path,
  bbox = NULL,
  calculate = TRUE,
  overwrite = FALSE,
  echo_cmd = FALSE,
  echo = TRUE,
  spinner = TRUE,
  verbose = FALSE,
  progress = FALSE
)

Arguments

input_path

String. Path to the existing OSM file.

output_path

String. Path where the new file will be written.

bbox

An sf::st_bbox to use. If NULL, the bounding box is computed via osm_get_bbox().

calculate

Logical(1). If TRUE, allow computing from data when no header is present; otherwise error if missing. Defaults to TRUE.

overwrite

Logical(1). If TRUE, allow overwriting output_path. Defaults to FALSE.

echo_cmd

A logical. Whether to print the Osmium command generated by the function call to the screen. Defaults to FALSE.

echo

A logical. Whether to print the standard output and error generated by the Osmium call to the screen. Defaults to TRUE.

spinner

A logical. Whether to show a reassuring spinner while the Osmium call is being executed. Defaults to TRUE.

verbose

A logical. Whether to display detailed information on the running command. Defaults to FALSE.

progress

A logical. Whether to display a progress bar while running the command. Defaults to FALSE.

Value

Invisibly returns the normalized path to output_path.

Examples

pbf <- system.file("extdata/cur.osm.pbf", package = "rosmium")

# Compute true bbox and write to a new file
out <- tempfile(fileext = ".osm.pbf")
osm_set_bbox(pbf, out, calculate = TRUE, overwrite = TRUE)
#> {
#>     "file": {
#>         "name": "/home/runner/work/_temp/Library/rosmium/extdata/cur.osm.pbf",
#>         "format": "PBF",
#>         "compression": "none",
#>         "size": 525406
#>     },
#>     "header": {
#>         "boxes": [
#>             [
#>                 -49.2888999,
#>                 -25.4412999,
#>                 -49.2509,
#>                 -25.4196999
#>             ]
#>         ],
#>         "with_history": false,
#>         "option": {
#>             "generator": "0.48.3",
#>             "pbf_dense_nodes": "true"
#>         }
#>     }
#> }

# Supply a custom bbox
library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
custom <- st_bbox(
  c(xmin = -49.3, ymin = -25.45, xmax = -49.24, ymax = -25.41),
  crs = st_crs(4326)
)
out2 <- tempfile(fileext = ".osm.pbf")
osm_set_bbox(pbf, out2, bbox = custom, overwrite = TRUE)